fix(acp): restore conversation flow and plan status
