+++ web/src/components/acp_panel.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { AcpView } from "../acp";
 import { AcpDebug, AcpDebugProps } from "./acp_debug";
 import { AcpConversation, AcpConversationProps } from "./acp_conversation";
-import { AcpPlan, AcpPlanProps } from "./acp_plan";
+import { AcpPlan, AcpPlanProps, summarizePlanEntries } from "./acp_plan";
 import {
   ACP_JUMP_BOTTOM_BUTTON_CLASS,
   ACP_PANEL_HEAD_CLASS,
@@ -51,6 +51,27 @@ function AcpPanelView({
   debug,
 }: AcpPanelProps) {
   const effectiveTab = !developerMode && acpTab === "debug" ? "conversation" : acpTab;
+  const planSummary = summarizePlanEntries(plan.plan?.entries ?? []);
+  const planStatus =
+    planSummary.active > 0
+      ? {
+          label: `${planSummary.active} active`,
+          className:
+            "bg-notion-accent/10 text-notion-accent",
+        }
+      : planSummary.pending > 0
+        ? {
+            label: `${planSummary.pending} pending`,
+            className:
+              "bg-amber-100 text-amber-800",
+          }
+        : planSummary.total > 0
+          ? {
+              label: "done",
+              className:
+                "bg-emerald-100 text-emerald-800",
+            }
+          : null;
   const tabButtonClassName = (selected: boolean, withGap = false) =>
     `${ACP_TAB_BUTTON_BASE_CLASS} ${withGap ? "gap-2 " : ""}${selected ? ACP_TAB_BUTTON_ACTIVE_CLASS : ACP_TAB_BUTTON_IDLE_CLASS}`;
   const tabsNode = (
@@ -73,6 +94,11 @@ function AcpPanelView({
         onClick={() => onSelectTab("plan")}
       >
         Plan
+        {planStatus ? (
+          <span className={`${ACP_TAB_BADGE_CLASS} ${planStatus.className}`}>
+            {planStatus.label}
+          </span>
+        ) : null}
       </button>
       {developerMode && (
         <button
