--- web/src/components/bubbles/thinking_bubble.tsx
@@ -1,82 +1,0 @@
-import React from "react";
-import { formatConversationPreview } from "../../conversation";
-import { ACP_BUBBLE_THINKING_CLASS } from "../../ui/tailwind_classes";
-import { ThreadRichText } from "../thread_rich_text";
-
-export type ThinkingBubbleProps = {
-  text: string;
-  live?: boolean;
-  summaryPrefix?: string;
-  grouped?: boolean;
-};
-
-export const ThinkingBubble = React.memo(function ThinkingBubble({
-  text,
-  live = false,
-  summaryPrefix,
-  grouped = false,
-}: ThinkingBubbleProps) {
-  const summary = deriveThinkingSummary(text, {
-    live,
-    summaryPrefix,
-  });
-  const entryClassName = grouped
-    ? ""
-    : " acp-row group relative flex w-full flex-col items-start px-4 py-1.5 sm:px-8";
-
-  return (
-    <div className={entryClassName}>
-      <div className={ACP_BUBBLE_THINKING_CLASS}>
-        <details className="acp-thought-fold acp-thinking-fold">
-          <summary className="cursor-pointer text-sm font-bold text-notion-text opacity-80">
-            {summary}
-          </summary>
-          <div className="acp-text mt-3 text-[14px] leading-relaxed text-notion-text opacity-90">
-            <ThreadRichText text={text} />
-          </div>
-        </details>
-      </div>
-    </div>
-  );
-});
-
-function deriveThinkingSummary(
-  text: string,
-  {
-    live,
-    summaryPrefix,
-  }: {
-    live?: boolean;
-    summaryPrefix?: string;
-  } = {}
-): string {
-  const normalizedText = unescapeLineBreaks(text);
-  const firstLine = normalizedText
-    .split("\n")
-    .map((line) => line.trim())
-    .find((line) => line.length > 0);
-  const preview = firstLine
-    ? formatConversationPreview(normalizeThinkingSummaryLine(firstLine), 96)
-    : "THINKING";
-  const prefix = summaryPrefix ?? "THINKING";
-  const base = preview === prefix ? prefix : `${prefix} · ${preview}`;
-  return live ? `${base} (live)` : base;
-}
-
-function normalizeThinkingSummaryLine(line: string): string {
-  const normalized = line
-    .replace(/!\[([^\]]*)\]\([^)]+\)/g, "$1")
-    .replace(/\[([^\]]+)\]\([^)]+\)/g, "$1")
-    .replace(/[`*_~>#]/g, "")
-    .replace(/\s+/g, " ")
-    .trim();
-  return normalized || line.trim();
-}
-
-function unescapeLineBreaks(text: string): string {
-  return text
-    .replace(/\\r\\n/g, "\n")
-    .replace(/\\n/g, "\n")
-    .replace(/\\t/g, "\t")
-    .replace(/\\r/g, "\n");
-}
